+++ Source/Graphics/Systems/TextureSystem.cpp
@@ -246,7 +246,7 @@ namespace shift::gfx {
                     continue;
                 }
 
-                ImGui::PushID(id);
+                ImGui::PushID(static_cast<int>(id));
 
                 if (ImGui::CollapsingHeader(std::string{name + "##" + std::to_string(id)}.c_str())) {
                     glm::ivec2 texSize = {m_system.m_textures[id]->GetWidth(), m_system.m_textures[id]->GetHeight()};
